Added electron configuration in element detail card
