+++ src/Components/PeriodicTable.css
@@ -5,6 +5,9 @@
   margin: 0 auto;
   overflow: visible;
 position: relative;
+display: flex;
+flex-direction: column;
+align-items: center;
 }
 
 /* ===== Controls Bar ===== */
@@ -83,7 +86,9 @@ position: relative;
   padding: 8px 0;
   margin-bottom: 12px;
   overflow: visible;
-position: relative;
+  position: relative;
+
+  justify-content: center;
 }
 
 /* ===== Element Cell ===== */
@@ -718,4 +723,44 @@ position: relative;
 model-viewer {
   --poster-color: transparent;
   background-color: transparent;
+}
+/* ===== MODAL ===== */
+
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.7);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 999;
+}
+
+.modal-content {
+  background: #111827;
+  color: white;
+  padding: 24px;
+  border-radius: 16px;
+  width: 400px;
+  max-width: 90%;
+  position: relative;
+  box-shadow: 0 0 30px rgba(0,0,0,0.5);
+}
+
+.close-modal {
+  position: absolute;
+  top: 10px;
+  right: 14px;
+  background: transparent;
+  border: none;
+  color: white;
+  font-size: 24px;
+  cursor: pointer;
+}
+
+.close-modal:hover {
+  color: #ff6b6b;
 }
